Combine cpp, move helpers to hpp (#2584)
